test(ledger): split v2 lifecycle integration tests and harden cancel cases :test_tube:
Move the commit/cancel lifecycle integration tests into their own file to stay under the size limit, add cancel unknown-id (404) and malformed-id (400) rows, assert the RFC 9457 code field exactly, and note the lifecycle terminals reuse the v1 shells.

X-Lerian-Ref: 0x1
